feat(about): enhance onboarding tour completion with polished dialog and correct navigation
