+++ src/lib/presentation/ui/shared/assets/locales/uk.yaml
@@ -91,7 +91,9 @@ shared:
     cannot_select_date_before_min_date: Не можна вибрати дату раніше {date}
     cannot_select_time_after_max_date: Не можна вибрати час пізніше {time}
     cannot_select_time_before_min_date: Не можна вибрати час раніше {time}
-    deadline_cannot_be_before_planned_date: Кінцевий термін має бути в день запланованої дати або пізніше
+    deadline_cannot_be_before_planned_date:
+      Кінцевий термін має бути в день запланованої
+      дати або пізніше
     end_date_cannot_be_after_max_date: Дата завершення не може бути пізніше {date}
     end_date_must_be_at_or_before: Дата завершення має бути не пізніше {date}
     select_date_range_title: Виберіть діапазон дат
@@ -99,7 +101,9 @@ shared:
     selected_date_must_be_at_or_after: Вибрана дата має бути не раніше {date}
     selected_date_must_be_at_or_before: Вибрана дата має бути не пізніше {date}
     selected_date_time_must_be_after: Вибрані дата і час мають бути пізніше {dateTime}
-    start_date_cannot_be_after_end_date: Дата початку не може бути пізніше за дату завершення
+    start_date_cannot_be_after_end_date:
+      Дата початку не може бути пізніше за дату
+      завершення
     start_date_cannot_be_before_min_date: Дата початку не може бути раніше {date}
     start_date_must_be_at_or_after: Дата початку має бути не раніше {date}
     time_must_be_at_or_after: Час має бути не раніше {time}
@@ -155,7 +159,8 @@ shared:
     loading: Сталася помилка під час завантаження даних
     report_subject: "{appName} App: Звіт про неочікувану помилку"
     report_template:
-      "Привіт, я зіткнувся з неочікуваною помилкою під час використання додатку {appName}.
+      "Привіт, я зіткнувся з неочікуваною помилкою під час використання
+      додатку {appName}.
 
 
       Ось інформація, яка може допомогти діагностувати проблему:
@@ -194,7 +199,9 @@ shared:
   startup_error:
     copied_to_clipboard: Деталі помилки скопійовано в буфер обміну
     copy_button: Копіювати деталі помилки
-    description: Щось пішло не так під час запуску додатка. Це може бути спричинено проблемою системи або несумісною конфігурацією.
+    description:
+      Щось пішло не так під час запуску додатка. Це може бути спричинено
+      проблемою системи або несумісною конфігурацією.
     details_title: Деталі Помилки
     report_button: Повідомити про Проблему
     title: Помилка Запуску Додатка
@@ -269,7 +276,9 @@ shared:
     average_daily_label: Середньоденний
     peak_hour_label: Пікова година
     select_date_range_message: Виберіть діапазон дат для фільтрації статистики
-    select_date_range_description: Виберіть власний діапазон дат для перегляду статистики використання
+    select_date_range_description:
+      Виберіть власний діапазон дат для перегляду статистики
+      використання
   time:
     not_set: Не встановлено
     minutes: хвилини
@@ -368,7 +377,9 @@ shared:
     time_must_be_at_or_before: Час має бути {time} або раніше
     selected_date_must_be_at_or_after: Вибрана дата має бути {date} або пізніше
     selected_date_must_be_at_or_before: Вибрана дата має бути {date} або раніше
-    start_date_cannot_be_after_end_date: Дата початку не може бути пізніше за дату завершення
+    start_date_cannot_be_after_end_date:
+      Дата початку не може бути пізніше за дату
+      завершення
     start_date_must_be_at_or_after: Дата початку має бути {date} або пізніше
     end_date_must_be_at_or_before: Дата завершення має бути {date} або раніше
     cannot_select_date_before_min_date: Не можна вибрати дату до {date}
@@ -426,3 +437,7 @@ shared:
     1_2_hours: 1 - 2 години
     more_than_2_hours: 2+ годин
   configure: Налаштувати
+  tour_finish_title: Екскурсія завершена
+  tour_finish_body:
+    Все готово! Удачі у вашому шляху до майстерності продуктивності.
+    Продовжуйте наполегливо працювати і насолоджуйтесь!
